+++ json_serializable/lib/src/decode_helper.dart
@@ -278,8 +278,7 @@ _ConstructorData _writeConstructorInvocation(
 
   for (final arg in ctor.parameters) {
     if (!availableConstructorParameters.contains(arg.name)) {
-      // ignore: deprecated_member_use
-      if (arg.isNotOptional) {
+      if (arg.isRequired) {
         var msg = 'Cannot populate the required constructor '
             'argument: ${arg.name}.';