+++ json_serializable/lib/src/type_helpers/json_converter_helper.dart
@@ -199,7 +199,7 @@ _JsonConvertData? _typeConverterFrom(
 
   final annotationElement = match.elementAnnotation?.element;
   if (annotationElement is PropertyAccessorElement) {
-    final enclosing = annotationElement.enclosingElement;
+    final enclosing = annotationElement.enclosingElement3;
 
     var accessString = annotationElement.name;
 
@@ -226,7 +226,7 @@ _JsonConvertData? _typeConverterFrom(
 
   if (match.genericTypeArg != null) {
     return _JsonConvertData.genericClass(
-      match.annotation.type!.element!.name!,
+      match.annotation.type!.element2!.name!,
       match.genericTypeArg!,
       reviver.accessor,
       match.jsonType,
@@ -235,7 +235,7 @@ _JsonConvertData? _typeConverterFrom(
   }
 
   return _JsonConvertData.className(
-    match.annotation.type!.element!.name!,
+    match.annotation.type!.element2!.name!,
     reviver.accessor,
     match.jsonType,
     match.fieldType,
@@ -263,18 +263,18 @@ _ConverterMatch? _compatibleMatch(
   ElementAnnotation? annotation,
   DartObject constantValue,
 ) {
-  final converterClassElement = constantValue.type!.element as ClassElement;
+  final converterClassElement = constantValue.type!.element2 as ClassElement;
 
   final jsonConverterSuper =
       converterClassElement.allSupertypes.singleWhereOrNull(
-    (e) => _jsonConverterChecker.isExactly(e.element),
+    (e) => _jsonConverterChecker.isExactly(e.element2),
   );
 
   if (jsonConverterSuper == null) {
     return null;
   }
 
-  assert(jsonConverterSuper.element.typeParameters.length == 2);
+  assert(jsonConverterSuper.element2.typeParameters.length == 2);
   assert(jsonConverterSuper.typeArguments.length == 2);
 
   final fieldType = jsonConverterSuper.typeArguments[0];
@@ -305,7 +305,7 @@ _ConverterMatch? _compatibleMatch(
       annotation,
       constantValue,
       jsonConverterSuper.typeArguments[1],
-      '${targetType.element.name}${targetType.isNullableType ? '?' : ''}',
+      '${targetType.element2.name}${targetType.isNullableType ? '?' : ''}',
       fieldType,
     );
   }