+++ json_annotation/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.7.0
+
+- Added `JsonEnum.valueField` which allows specifying a field in an 
+  "enhanced enum" to use for serialization instead of specifying each value
+  individually with `JsonValue`.
+- Require Dart SDK 2.17
+
 ## 4.6.0
 
 - Added `JsonSerializable(createFieldMap: true)`.